+++ src/titiler/core/titiler/core/__init__.py
@@ -1,11 +1,14 @@
 """titiler.core"""
 
-__version__ = "0.18.6"
+__version__ = "0.18.9"
 
 from . import dependencies, errors, factory, routing  # noqa
 from .factory import (  # noqa
-    BaseTilerFactory,
+    AlgorithmFactory,
+    BaseFactory,
+    ColorMapFactory,
     MultiBandTilerFactory,
     MultiBaseTilerFactory,
     TilerFactory,
+    TMSFactory,
 )