+++ src/titiler/extensions/titiler/extensions/__init__.py
@@ -1,6 +1,6 @@
 """titiler.extensions"""
 
-__version__ = "0.18.6"
+__version__ = "0.18.9"
 
 from .cogeo import cogValidateExtension  # noqa
 from .stac import stacExtension  # noqa